Properly consolidate error handling implementations (Issue #189)
Integrate the factory pattern directly into the existing error handling system:

- Add factory pattern to error_handled_base.py for DI compatibility
- Update error_handled_article.py with injectable provider
- Update articles.py to use the extended CRUD implementation
- Remove parallel implementation files
